added multer config in a separate file
